requirements.txt
streamlit==1.31.0
chromadb==0.4.22
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
python-dotenv==1.0.0
sentence-transformers==2.5.1
groq==0.4.2
